+++ src/components/SelectedFilter.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import Chip from '@mui/material/Chip'
 import { useDispatch, useSelector } from 'react-redux'
 import { makeStyles } from '@mui/styles'
-import { useHistory } from 'react-router-dom'
+import { useNavigate, useLocation, useSearchParams } from 'react-router-dom'
 
 import { filterSelector } from 'selectors/filterSelectors'
 import { filterRemoved } from 'actions/galleryActions'
@@ -16,15 +16,21 @@ const useStyles = makeStyles((theme) => ({
 
 export default function SelectedFilter({ galleryId, filterId }) {
   const classes = useStyles()
-  const history = useHistory()
+  const location = useLocation()
+  const navigate = useNavigate()
+  const [searchParams] = useSearchParams()
   const dispatch = useDispatch()
   const filter = useSelector((state) => filterSelector(state, { filterId }))
 
   const onDelete = useCallback(() => {
-    dispatch(filterRemoved(galleryId, filterId, history))
-  }, [galleryId, filterId])
+    dispatch(filterRemoved(galleryId, filterId, navigate, location, searchParams))
+  }, [galleryId, filterId, navigate, location, searchParams])
 
-  return <Chip className={classes.filterItem} color="primary" label={filter.name} onDelete={onDelete} />
+  return (
+    <span className={classes.filterItem}>
+      <Chip color="primary" label={filter.name} onDelete={onDelete} />
+    </span>
+  )
 }
 
 SelectedFilter.propTypes = {